Wrapper: fallback radspec evaluation to check all installed apps (#376)
